requirements.txt
pytesseract
Pillow
openai
Faker
nltk
selenium
bs4
opencv-python
poe-api